REFACTOR: Renamed new AdvancedEditorAreaPane to SplitEditorAreaPane.
The previous implementation is available under the name AdvancedEditorAreaPane.
