+++ pyface/tasks/api.py
@@ -1,5 +1,6 @@
 # Local imports.
 from advanced_editor_area_pane import AdvancedEditorAreaPane
+from split_editor_area_pane import SplitEditorAreaPane
 from dock_pane import DockPane
 from editor import Editor
 from editor_area_pane import EditorAreaPane
