+++ pyface/tasks/split_editor_area_pane.py
@@ -1,0 +1,4 @@
+# Import the toolkit specific version.
+from pyface.toolkit import toolkit_object
+SplitEditorAreaPane = toolkit_object('tasks.split_editor_area_pane:'
+                                     'SplitEditorAreaPane')
